% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfarc.node-package.R
\docType{package}
\name{sfarc.node-package}
\alias{sfarc.node}
\alias{sfarc.node-package}
\title{sfarc.node: Create Lines from Shared Boundaries}
\description{
Create lines from shared boundaries between polygons. A single copy of a shared boundary is 
 created as 'sf' lines, keeping the row identity of the input polygon/s. It's expected that polygons have
 clean neighbour boundaries and don't overlap.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
